requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==2.0.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
charset-normalizer==3.1.0
Discord-Extension-Interaction @ git+https://github.com/gunyu1019/discord-extension-interaction@114df968a5bca9cc94a95ef471a84025c3b0eabb
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
yarl==1.8.2
